requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
flask>=2.0.0
flask-cors>=3.0.0
pillow>=9.0.0
numpy>=1.21.0
pandas>=1.3.0
plotly>=5.0.0
